Pull search algorithm out into another module
